\name{plot.partynat}
\alias{plot.partynat}
\title{
	function for plotting objects of class \code{'partynat'}
}
\description{
	The function produces a dot-plot of idnex values for objects
	of class \code{'partynat'}. 
}
\usage{
plot.partynat(x, ...)
}
\arguments{
  \item{x}{
	  An object of S3 class \code{'partynat'}.
}
  \item{\dots}{
	  Further arguments to be passed to \code{plot()}.
}
}
\details{
}
\value{
	A plot.
}
\author{
	Juraj Medzihorsky
}
\seealso{
	\code{\link[partynat::partynat]{partynat}}
}
\examples{}
\keyword{ partynat }
\keyword{ plot }
