\name{partynat-package}
\alias{partynat-package}
\docType{package}
\title{
	partynat, a package for party nationalization indices
}
\description{
	partynat is a package for indices of party nationalization and related
	quantities.
}
\details{
\tabular{ll}{
Package: \tab partynat\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2022-05-17\cr
License: \tab GPL >= 2\cr
}
}
\author{
Juraj Medzihorsky <juraj.medzihorsky@gmail.com>
}
\references{
	Medzihorsky, Juraj. 2022. Unifying the measurement of variation in electoral support.
}
\keyword{ package }
\keyword{ nationalization }
\keyword{ resampling }
\seealso{
}
\examples{
}
